// src/app/enums/trade-types.ts
export enum TradeType {
    plumber = 'Plumber',
    electrician = 'Electrician',
    painter = 'Painter/Decorator',
    plaster = 'Plastering',
    locksmith = 'Locksmith',
    builder = 'Builder',
    cleaner = 'Cleaner',
    carpentry = 'Carpentry',
    landscaper = 'Landscaper',
    other = 'Other'

}